use CDI events for delivering to JMS queues to simplify split per networks
