--- push/sender/src/main/java/org/jboss/aerogear/unifiedpush/message/SenderMDB.java
@@ -1,129 +1,0 @@
-package org.jboss.aerogear.unifiedpush.message;
-
-import java.util.List;
-
-import javax.annotation.Resource;
-import javax.ejb.ActivationConfigProperty;
-import javax.ejb.MessageDriven;
-import javax.ejb.TransactionAttribute;
-import javax.ejb.TransactionAttributeType;
-import javax.enterprise.inject.Any;
-import javax.enterprise.inject.Instance;
-import javax.inject.Inject;
-import javax.jms.Connection;
-import javax.jms.ConnectionFactory;
-import javax.jms.JMSException;
-import javax.jms.MessageListener;
-import javax.jms.MessageProducer;
-import javax.jms.ObjectMessage;
-import javax.jms.Queue;
-import javax.jms.Session;
-
-import org.jboss.aerogear.unifiedpush.api.PushMessageInformation;
-import org.jboss.aerogear.unifiedpush.api.Variant;
-import org.jboss.aerogear.unifiedpush.api.VariantMetricInformation;
-import org.jboss.aerogear.unifiedpush.message.sender.NotificationSenderCallback;
-import org.jboss.aerogear.unifiedpush.message.sender.PushNotificationSender;
-import org.jboss.aerogear.unifiedpush.message.sender.SenderTypeLiteral;
-import org.jboss.aerogear.unifiedpush.utils.AeroGearLogger;
-
-@MessageDriven(name = "TokenBatchQueue", activationConfig = {
-        @ActivationConfigProperty(propertyName = "destination", propertyValue = "queue/TokenBatchQueue"),
-        @ActivationConfigProperty(propertyName = "destinationType", propertyValue = "javax.jms.Queue"),
-        @ActivationConfigProperty(propertyName = "acknowledgeMode", propertyValue = "Auto-acknowledge") })
-public class SenderMDB implements MessageListener {
-
-    private final AeroGearLogger logger = AeroGearLogger.getInstance(SenderMDB.class);
-
-    @Inject
-    @Any
-    private Instance<PushNotificationSender> senders;
-
-    @Resource(mappedName = "java:/ConnectionFactory")
-    private ConnectionFactory connectionFactory;
-
-    @Resource(mappedName = "java:/queue/VariantMetricQueue")
-    private Queue metricsQueue;
-
-    @TransactionAttribute(TransactionAttributeType.NOT_SUPPORTED)
-    public void onMessage(javax.jms.Message jmsMessage) {
-        try {
-            if (jmsMessage instanceof ObjectMessage) {
-                final MessageWithTokens msg = (MessageWithTokens) ((ObjectMessage) jmsMessage).getObject();
-                final Variant variant = msg.getVariant();
-                final UnifiedPushMessage message = msg.getUnifiedPushMessage();
-                final List<String> deviceTokens = msg.getDeviceTokens();
-
-                logger.fine("Received message from queue: " + msg.getUnifiedPushMessage().getMessage().getAlert());
-
-                senders.select(new SenderTypeLiteral(variant.getClass())).get()
-                                    .sendPushMessage(variant, deviceTokens, message, new SenderServiceCallback(variant, deviceTokens.size(), msg.getPushMessageInformation()));
-            } else {
-                logger.warning("Received message of wrong type: " + jmsMessage.getClass().getName());
-            }
-        } catch (JMSException e) {
-            throw new MessageDeliveryException("Failed to handle message from TokenBatchQueue", e);
-        }
-    }
-
-    private class SenderServiceCallback implements NotificationSenderCallback {
-        private final Variant variant;
-        private final int tokenSize;
-        private final PushMessageInformation pushMessageInformation;
-
-        public SenderServiceCallback(Variant variant, int tokenSize, PushMessageInformation pushMessageInformation) {
-            this.variant = variant;
-            this.tokenSize = tokenSize;
-            this.pushMessageInformation = pushMessageInformation;
-        }
-
-        @Override
-        public void onSuccess() {
-            logger.fine(String.format("Sent '%s' message to '%d' devices", variant.getType().getTypeName(), tokenSize));
-            updateStatusOfPushMessageInformation(pushMessageInformation, variant.getVariantID(), tokenSize, Boolean.TRUE);
-        }
-
-        @Override
-        public void onError(final String reason) {
-            logger.warning(String.format("Error on '%s' delivery", variant.getType().getTypeName()));
-            updateStatusOfPushMessageInformation(pushMessageInformation, variant.getVariantID(), tokenSize, Boolean.FALSE, reason);
-        }
-    }
-
-    /**
-     * Helpers to update the given {@link PushMessageInformation} with a {@link VariantMetricInformation} object
-     */
-    private void updateStatusOfPushMessageInformation(final PushMessageInformation pushMessageInformation, final String variantID, final int receivers, final Boolean deliveryStatus) {
-        this.updateStatusOfPushMessageInformation(pushMessageInformation, variantID, receivers, deliveryStatus, null);
-    }
-
-    private void updateStatusOfPushMessageInformation(final PushMessageInformation pushMessageInformation, final String variantID, final int receivers, final Boolean deliveryStatus, final String reason) {
-        final VariantMetricInformation variantMetricInformation = new VariantMetricInformation();
-        variantMetricInformation.setPushMessageInformation(pushMessageInformation);
-        variantMetricInformation.setVariantID(variantID);
-        variantMetricInformation.setReceivers(receivers);
-        variantMetricInformation.setDeliveryStatus(deliveryStatus);
-        variantMetricInformation.setReason(reason);
-
-        Connection connection = null;
-        try {
-            connection = connectionFactory.createConnection();
-            Session session = connection.createSession(false, Session.AUTO_ACKNOWLEDGE);
-            MessageProducer messageProducer = session.createProducer(metricsQueue);
-            connection.start();
-
-            ObjectMessage messageWithMetrics = session.createObjectMessage(variantMetricInformation);
-            messageProducer.send(messageWithMetrics);
-        } catch (JMSException e) {
-            throw new MessageDeliveryException("Failed to queue variant metrics for further processing", e);
-        } finally {
-            if (connection != null) {
-                try {
-                    connection.close();
-                } catch (JMSException e) {
-                    e.printStackTrace();
-                }
-            }
-        }
-    }
-}
