+++ push/sender/src/main/java/org/jboss/aerogear/unifiedpush/message/jms/Dequeue.java
@@ -1,0 +1,14 @@
+package org.jboss.aerogear.unifiedpush.message.jms;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+import javax.inject.Qualifier;
+
+@Qualifier
+@Target({ ElementType.PARAMETER, ElementType.FIELD })
+@Retention(RetentionPolicy.RUNTIME)
+public @interface Dequeue {
+}
